Task Ultima | INNOSHOP | Fluent Validation Update
Added:
Fluent Validation rules for Registration, Login and Products.
Global Exception Handler for all Exceptions.

In test (WIP):
Updated Guard system (Bussines Logic Validators).

+++ Inno_Shop_Cherkasov/InnoShop.Core.Services/Exceptions/NegativeNumberException.cs
@@ -3,8 +3,8 @@
 {
     public class NegativeNumberException : Exception
     {
-        public int Number { get; }
-        public NegativeNumberException(int number, string message) : base($"{number}, {message}")
+        public decimal Number { get; }
+        public NegativeNumberException(decimal number, string message) : base($"{number}, {message}")
         {
             this.Number = number;
         }
